% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evallib.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(confmat)
}
\arguments{
\item{confmat}{A confusion matrix}
}
\value{
The accuracy value of the predicted and actual values
}
\description{
accuracy
}
