% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{actAssertions}
\alias{actAssertions}
\title{actAssertions}
\usage{
actAssertions(x)
}
\arguments{
\item{x}{A vector of actual cases}
}
\value{
TRUE or FALSE depending on the result
}
\description{
Assert things about cases
}
\examples{
tdat
actAssertions(tdat$actual)
}
