% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{auc}
\alias{auc}
\title{auc}
\usage{
auc(x, y)
}
\arguments{
\item{x}{The X dimension of the ROC curve; Fallout scores.}

\item{y}{The Y dimension of the ROC curve; Recall scores.}
}
\value{
An AUC score
}
\description{
Returns the Area Under Curve statistic, which is often used to interpret an ROC curve.
}
\examples{
predicted <- sample(c(1,0), replace = TRUE, size = 150)
actual <- sample(c(1,0), replace = TRUE, size = 150)
myRoc <- roc(predicted,actual)
auc(myRoc$fallout, myRoc$recall)
}
