% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aucWithCI}
\alias{aucWithCI}
\title{aucWithCI}
\usage{
aucWithCI(pred, actual)
}
\arguments{
\item{pred}{Vector of predictions}

\item{actual}{Vector of outcomes}
}
\value{
A list with score and quantiles
}
\description{
Wrapper for aucFromPA that also calls \code{pROC::ci.auc} to obtain DeLong
confidence intervals.
}
\examples{
testdata <- examplePredictions() 
aucWithCI(testdata$pred, testdata$actual) 

}
