% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(pred, actual, fun, draws = 100, parallel = FALSE, ...)
}
\arguments{
\item{pred}{A vector of predictions}

\item{actual}{A vector of actual values}

\item{fun}{The function to apply}

\item{draws}{Number of times to randomly sample the data for bootstrapping.
Larger numbers yield better results, but also increase runtime.}

\item{parallel}{Bool. Run in parallel?}

\item{...}{Arguments passed to fun}
}
\value{
A list containing the results of each iteration
}
\description{
Apply \code{fun} to the data \code{n} times.  The data is randomly sampled
each time. This yields both more robust results and approximate confidence
intervals for the results.  Due to the random sampling, the results of this
function are stochastic. Set the seed with \code{set.seed(n)} for
reproducible results.
}
\examples{
pred <- sample(seq(0,1,0.001), size = 150, replace = TRUE)
actual <- sample(c(0,1), size = 150, replace = TRUE)
bootstrap(pred,actual,roc)

}
