% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrappedMetricCurve}
\alias{bootstrappedMetricCurve}
\title{bootstrappedMetricCurve}
\usage{
bootstrappedMetricCurve(pred, actual, x, y, res = 0.1, draws = 100,
  parallel = FALSE)
}
\arguments{
\item{pred}{Probability predictions}

\item{actual}{Outcomes}

\item{res}{Resolution of ROC thresholds. A float between 0 and 1.}

\item{draws}{Number of times to randomly sample the data for bootstrapping.
Larger numbers yield better results, but also increase runtime.}

\item{parallel}{Boolean. Run in parallel? Uses ncores - 1 CPU cores.}

\item{probs}{Vector of quantiles to return}
}
\value{
A list containing ROC data and AUC with confidence intervals.
}
\description{
Makes \code{metricCurve} data using bootstrapping, which makes it  
possible to calculate the standard deviation, and the quantiles of 
the results.
}
\details{
Note that the function requires specification of the resolution of the
\code{metricCurve}. This is because each bootstrap iteration must have the
same dimensions, since it must be possible to collapse these into an array.
}
\examples{
tdat <- examplePredictions()
bootstrappedMetricCurve(tdat$pred,tdat$actual,roc)

}
