% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrappedROC}
\alias{bootstrappedROC}
\title{bootstrappedROC}
\usage{
bootstrappedROC(pred, actual, ...)
}
\arguments{
\item{pred}{Probability predictions}

\item{actual}{Outcomes}
}
\value{
A list containing ROC data and AUC with confidence intervals.
}
\description{
Helper function that makes ROC graph data from predictions and outcomes with
confidence intervals and standard deviation.
}
\examples{
pred <- sample(seq(0,1,0.001), size = 150, replace = TRUE)
actual <- sample(c(0,1), size = 150, replace = TRUE)
bootstrappedROC(pred,actual,roc)

}
