% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{categorizeClasf}
\alias{categorizeClasf}
\title{categorizeClasf}
\usage{
categorizeClasf(pred, act)
}
\arguments{
\item{pred}{Predicted values}

\item{pred}{Actual values}
}
\value{
A character vector
}
\description{
Just names a combination of predicted / actual values as either TP, FP, TN, FN.
}
