% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cintervalplot}
\alias{cintervalplot}
\title{cintervalplot}
\usage{
cintervalplot(pred, actual, res = 0.1, parallel = TRUE)
}
\arguments{
\item{pred}{Vector of predictions}

\item{actual}{Vector of outcomes}

\item{...}{Parameters passed on to \code{bootstrappedROC}}
}
\value{
A ggplot object
}
\description{
Just a helpful wrapper that demonstrates how to make a confidence-interval
plot from raw probability-outcome data.
}
\examples{
testdata <- examplePredictions() 
cintervalplot(testdata$pred, testdata$actual)

}
