% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cintervalplot2}
\alias{cintervalplot2}
\title{cintervalplot2}
\usage{
cintervalplot2(pred, actual, x, y, draws = 100, res = 0.1,
  parallel = FALSE)
}
\arguments{
\item{pred}{Vector of predictions}

\item{actual}{Vector of outcomes}

\item{x}{The X dimension of the metric curve.}

\item{y}{The Y dimension of the metric curve.}

\item{draws}{Number of times to randomly sample the data for bootstrapping.
Larger numbers yield better results, but also increase runtime.}

\item{res}{Resolution of ROC thresholds. A float between 0 and 1.}

\item{parallel}{Boolean. Run in parallel? Uses ncores - 1 CPU cores.}
}
\value{
A list containing results and a ggplot object.
}
\description{
Just a helpful wrapper that demonstrates how to make a confidence-interval
plot from raw probability-outcome data, using metric functions x and y.
}
\examples{
# To make a ROC plot with confidence intervals: 
testdata <- examplePredictions() 
cintervalplot(testdata$pred, testdata$actual, fallout, recall,
              draws = 1000, res = 0.01, parallel = TRUE)

}
