% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{confmat}
\alias{confmat}
\title{bootstrap
confmat}
\usage{
confmat(predicted, actual)
}
\arguments{
\item{predicted}{Predicted values}

\item{actual}{Actual values}
}
\value{
A 2x2 matrix with TN,FN,FP,TP
}
\description{
Returns a 2x2 matrix with the results of a classification procedure. Useful
for computing further statistics.
}
\examples{
confmat(c(1,1,0),c(0,1,0))


}
