% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{examplePredictions}
\alias{examplePredictions}
\title{examplePredictions}
\usage{
examplePredictions(n = 150, noise = 0.2)
}
\value{
A data.frame with $pred and $actual columns.
}
\description{
Generates two vectors, one with predictions and one with outcomes.  Used to
test functions. Stochastic, seed must be set to get reproducible results.
}
\examples{
set.seed(1337)
tdat <- examplePredictions()
bootstrap(tdat$pred, tdat$actual, roc)
}
