% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{fscore}
\alias{fscore}
\title{fscore}
\usage{
fscore(precision, recall, beta = 1)
}
\arguments{
\item{precision}{Precision statistic}

\item{recall}{Recall statistic}
}
\value{
F-statistic
}
\description{
The F-score is computed from precision and recall, representing a balance
between the two statistics.
}
