% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metricCurve}
\alias{metricCurve}
\title{metricCurve}
\usage{
metricCurve(p, actual, x, y, res = NULL)
}
\arguments{
\item{p}{Vector of predicted probabilities}

\item{actual}{Vector of actual scores}

\item{x}{A function operating on the 2x2 confusion matrix, yielding a score.
Common functions are precision, recall, fallout and accuracy.}

\item{y}{A function operating on the 2x2 confusion matrix, yielding a score.
Common functions are precision, recall, fallout and accuracy.}

\item{res}{Specification of the threshold resolution of the curve. Either a
single number, a vector of numbers, or NULL. If a single number, it will be
used as an interval to create thresholds between 0 and 1. If a vector, it
will be used  as the vector of thresholds. If NULL, thresholds will be based
on the unique values of p.}
}
\value{
A data-frame of metric scores, that can be used to create an
interesting plot, or to calculate the AUC.
}
\description{
Creates a Metric Curve dataset, that can be used to create a metric plot,
like a ROC plot or a Precision-recall plot, and calculate AUC. ROC is a
special, widely used case of the MC.
}
\examples{
predicted <- sample(seq(0,1,0.01),replace = TRUE, size = 150)
actual <- sample(c(1,0),replace = TRUE, size = 150)
metricCurve(predicted, actual, precision, recall)
}
