% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{precision}
\alias{precision}
\title{precision}
\usage{
precision(confmat)
}
\arguments{
\item{confmat}{A confusion matrix}
}
\value{
The precision value of the predicted and actual values
}
\description{
precision
}
