% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{predAssertions}
\alias{predAssertions}
\title{predAssertions}
\usage{
predAssertions(x)
}
\arguments{
\item{x}{A vector of predictions}
}
\value{
TRUE or FALSE depending on the result
}
\description{
Assert things about predictions
}
\examples{
tdat
predAssertions(tdat$pred)
}
