% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{roc}
\alias{roc}
\title{roc}
\usage{
roc(p, actual, res = 0.01)
}
\arguments{
\item{p}{Vector of predicted probabilities}

\item{actual}{Vector of actual scores}
}
\value{
A data-frame of ROC scores.
}
\description{
Creates a data-frame that can be used to plot a ROC curve, or calculate the AUC.
Each row contains a fallout (x) and recall (y) score, that is calculated
with a given threshold for the predicted probabilities, which determines
which predicted values are interpreted as actual positives.
}
\examples{
predicted <- sample(seq(0,1,0.01),replace = TRUE, size = 150)
actual <- sample(c(1,0),replace = TRUE, size = 150)
roc(predicted,actual)
}
