% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evallib.R
\name{withConfmat}
\alias{withConfmat}
\title{withConfmat}
\usage{
withConfmat(p, a, fun)
}
\arguments{
\item{p}{Predicted values}

\item{a}{Actual values}
}
\value{
The result of whatever function is passed as fun (recall/precision/accuracy etc.)
}
\description{
Helper function. Computes a statistic from predicted and actual values,
using an intermediate confusion matrix.
}
\examples{
battery <- list(precision = precision, recall = recall,
                fallout = fallout, accuracy = accuracy) 
pred <- sample(c(1,0),replace = TRUE, size = 150)
actual <- sample(c(1,0),replace = TRUE, size = 150)
lapply(battery,function(fun){
       withConfmat(pred,actual,fun)})
}
