% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{zerosafeDiv}
\alias{zerosafeDiv}
\title{zerosafeDiv}
\usage{
zerosafeDiv(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
A number
}
\description{
Division, with the exception that division by zero yields one.
}
\examples{
zerosafeDiv(1,0) 
[1] 1
zerosafeDiv(1,1) 
[1] 1
}
